--- packages/@studiocms/oauth/src/github/effect.ts
@@ -1,175 +1,0 @@
-import { getSecret } from 'astro:env/server';
-import { Session, User, VerifyEmail } from 'studiocms:auth/lib';
-import { StudioCMSRoutes } from 'studiocms:lib';
-import { SDKCore } from 'studiocms:sdk';
-import { GitHub, generateState } from 'arctic';
-import type { APIContext } from 'astro';
-import { Effect, genLogger, Platform, Schema } from 'studiocms/effect';
-import { getCookie, getUrlParam, ValidateAuthCodeError } from 'studiocms/oAuthUtils';
-import {
-	handleExistingOAuthAccount,
-	handleNewOAuthUser,
-	handleOAuthLinking,
-} from '../shared.js';
-
-/**
- * Represents a GitHub user profile as returned by the GitHub API.
- *
- * @property id - The unique identifier for the user.
- * @property html_url - The URL to the user's GitHub profile.
- * @property login - The user's GitHub username.
- * @property avatar_url - The URL to the user's avatar image.
- * @property name - The user's display name.
- * @property blog - The user's blog URL.
- * @property email - The user's public email address.
- */
-export class GitHubUser extends Schema.Class<GitHubUser>('GitHubUser')({
-	id: Schema.Number,
-	html_url: Schema.String,
-	login: Schema.String,
-	avatar_url: Schema.String,
-	name: Schema.optional(Schema.String),
-	blog: Schema.optional(Schema.String),
-	email: Schema.optional(Schema.String),
-}) {}
-
-const GITHUB = {
-	CLIENT_ID: getSecret('CMS_GITHUB_CLIENT_ID') ?? '',
-	CLIENT_SECRET: getSecret('CMS_GITHUB_CLIENT_SECRET') ?? '',
-	REDIRECT_URI: getSecret('CMS_GITHUB_REDIRECT_URI') ?? null,
-};
-
-/**
- * Provides GitHub OAuth authentication effects for the StudioCMS API.
- *
- * This service handles the OAuth flow with GitHub, including:
- * - Initializing the OAuth session and redirecting users to GitHub for authorization.
- * - Validating the authorization code returned by GitHub and fetching user data.
- * - Handling the callback from GitHub, including:
- *   - Linking OAuth accounts to existing users.
- *   - Creating new users from GitHub profile data.
- *   - Verifying user email addresses.
- *   - Creating user sessions and redirecting to appropriate pages.
- *
- * @remarks
- * - Depends on session management, user library, email verification, and SDK core services.
- * - Handles both first-time setup and linking additional OAuth providers to existing accounts.
- * - Note: GitHub does not use PKCE (no code verifier).
- *
- * @example
- * ```typescript
- * const githubOAuth = new GitHubOAuthAPI();
- * yield* githubOAuth.initSession(context);
- * yield* githubOAuth.initCallback(context);
- * ```
- *
- * @see {@link Session}
- * @see {@link SDKCore}
- * @see {@link VerifyEmail}
- * @see {@link User}
- */
-export class GitHubOAuthAPI extends Effect.Service<GitHubOAuthAPI>()('GitHubOAuthAPI', {
-	dependencies: [VerifyEmail.Default, Platform.FetchHttpClient.layer],
-	effect: genLogger('studiocms/routes/api/auth/github/effect')(function* () {
-		const [sdk, fetchClient, { setOAuthSessionTokenCookie }, { createOAuthUser }] =
-			yield* Effect.all([SDKCore, Platform.HttpClient.HttpClient, Session, User]);
-
-		const { CLIENT_ID, CLIENT_SECRET, REDIRECT_URI } = GITHUB;
-
-		const github = new GitHub(CLIENT_ID, CLIENT_SECRET, REDIRECT_URI);
-
-		const validateAuthCode = (code: string) =>
-			genLogger('studiocms/routes/api/auth/github/effect.validateAuthCode')(function* () {
-				const tokens = yield* Effect.tryPromise(() => github.validateAuthorizationCode(code));
-
-				return yield* fetchClient
-					.get('https://api.github.com/user', {
-						headers: {
-							Authorization: `Bearer ${tokens.accessToken}`,
-						},
-					})
-					.pipe(
-						Effect.flatMap(Platform.HttpClientResponse.schemaBodyJson(GitHubUser)),
-						Effect.mapError(
-							(error) =>
-								new ValidateAuthCodeError({
-									provider: GitHubOAuthAPI.ProviderID,
-									message: `Failed to fetch user info: ${error.message}`,
-								})
-						)
-					);
-			});
-
-		return {
-			initSession: (context: APIContext) =>
-				genLogger('studiocms/routes/api/auth/github/effect.initSession')(function* () {
-					const state = generateState();
-					const scopes = ['user:email', 'repo'];
-
-					const url = github.createAuthorizationURL(state, scopes);
-
-					yield* setOAuthSessionTokenCookie(context, GitHubOAuthAPI.ProviderCookieName, state);
-
-					return context.redirect(url.toString());
-				}),
-			initCallback: (context: APIContext) =>
-				genLogger('studiocms/routes/api/auth/github/effect.initCallback')(function* () {
-					const { redirect } = context;
-
-					const [code, state, storedState] = yield* Effect.all([
-						getUrlParam(context, 'code'),
-						getUrlParam(context, 'state'),
-						getCookie(context, GitHubOAuthAPI.ProviderCookieName),
-					]);
-
-					if (!code || !state || !storedState || state !== storedState) {
-						return redirect(StudioCMSRoutes.authLinks.loginURL);
-					}
-
-					const githubUser = yield* validateAuthCode(code);
-
-					// GitHub returns a numeric id; coerce to string for the OAuth provider API.
-					const { id: githubUserId, login: githubUsername } = githubUser;
-					const githubUserIdStr = `${githubUserId}`;
-
-					const existingOAuthAccount = yield* sdk.AUTH.oAuth.searchProvidersForId({
-						providerId: GitHubOAuthAPI.ProviderID,
-						userId: githubUserIdStr,
-					});
-
-					if (existingOAuthAccount) {
-						return yield* handleExistingOAuthAccount(context, existingOAuthAccount).pipe(Effect.provide(VerifyEmail.Default));
-					}
-
-					const linkResult = yield* handleOAuthLinking(
-						context,
-						GitHubOAuthAPI.ProviderID,
-						githubUserIdStr
-					).pipe(Effect.provide(VerifyEmail.Default));
-					if (linkResult) return linkResult;
-
-					const newUser = yield* createOAuthUser(
-						{
-							id: crypto.randomUUID(),
-							username: githubUsername,
-							email: githubUser.email || null,
-							name: githubUser.name || githubUsername,
-							avatar: githubUser.avatar_url,
-							createdAt: new Date().toISOString(),
-							url: githubUser.blog || null,
-							emailVerified: false,
-							notifications: null,
-							password: null,
-							updatedAt: new Date().toISOString(),
-						},
-						{ provider: GitHubOAuthAPI.ProviderID, providerUserId: githubUserIdStr }
-					);
-
-					return yield* handleNewOAuthUser(context, newUser).pipe(Effect.provide(VerifyEmail.Default));
-				}),
-		};
-	}),
-}) {
-	static ProviderID = 'github';
-	static ProviderCookieName = 'github_oauth_state';
-}
